Decorator refactor
