+++ impl/core/src/main/java/io/serverlessworkflow/impl/WorkflowDecorator.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2020-Present The Serverless Workflow Specification Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.serverlessworkflow.impl;
+
+/**
+ * Parent interface for decorating object requests with additional modifications.<br>
+ * It is expected that this class is inherited by specializations that indicate the type to be
+ * decorated.
+ *
+ * <p>Children interfaces should be loaded via ServiceLoader and are sorted by priority in ascending
+ * order (lower priority numbers executed first).
+ *
+ * @see ServicePriority
+ */
+public interface WorkflowDecorator<T> extends ServicePriority {
+
+  /**
+   * Decorates an object with information extracted from the workflow and task context
+   *
+   * @param decorated
+   * @param workflowContext
+   * @param taskContext
+   */
+  void decorate(T decorated, WorkflowContext workflowContext, TaskContext taskContext);
+}
